Adding CANSniffer to GUI
A Teensy 4.0 is connected to the CAN bus and listens to all bus traffic, then publishes the number of messages received over the nominated interval back to the CAN bus. This message is then taken (via can_handler) to ROS2, where it can be plotted in the GUI. Very useful for debugging!