Major revision: Run the Ansible tasks in the Dockerfile
Best practice for Docker is to build everything in the Dockerfile rather
than to commit changes to a running container because committing a
container (a) creates large images and (b) doesn't clearly document the
build process.

To make this change, we need to make base_setup.sh noninteractive when
being run by the Dockerfile. This involved moving the construction of
the ansible-playbook command into the setup.sh for each environment and
then passing that command to base_setup.sh. This reorganization has the
additional benefit that flags to the setup.sh scripts are no longer
coddled, which makes these scripts much more readable.