Major revision: Run the docker container as a service
Previously, we were creating at transient container for testing. Now,
we're creating a named container and then starting and stopping it using
a systemd unit. The docker-native-env script now starts a new terminal
session inside the container using `docker exec`.

We also introduce a named volume for the /home directory in the
container.

This revision gives responsibility to the WSL2 env provisioning
for installing the management facilities for the guest docker
environment. Previously, some of that had been in the docker native
provisioning scripts.

The Dockerfile now has task documentation.